+++ Sources/RichTextKit/Commands/RichTextFormatCommandMenu.swift
@@ -1,5 +1,5 @@
 //
-//  RichTextFormatCommands.swift
+//  RichTextFormatCommandMenu.swift
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2022-12-20.
@@ -9,11 +9,23 @@
 import SwiftUI
 
 /**
- This Commands type can be used to expose a rich text format
- menu to an app's main menu or key commands.
- */
-public struct RichTextFormatCommands: Commands {
+ This command menu can be used to add rich text menu options
+ to an app, to control rich text formatting.
 
+ The commands will be disabled if there's no focus value for
+ a ``RichTextContext`` available. Make sure to add it to the
+ editor to make these commands work:
+
+ ```swift
+ RichTextEditor(...)
+    .focusedValue(\.richTextContext, richTextContext)
+ ```
+ */
+public struct RichTextFormatCommandMenu: Commands {
+
+    /**
+     Create a rich text format command menu.
+     */
     public init() {}
 
     @FocusedValue(\.richTextContext)