+++ Sources/RichTextKit/Localization/RTKL10n.swift
@@ -41,6 +41,7 @@ public enum RTKL10n: String, CaseIterable, Identifiable {
     menuFontLarger,
     menuFontSmaller,
     menuFormat,
+    menuPrint,
     menuSave,
     menuSaveAs,
     menuShare,